Enum kanidm_cli::ServiceAccountApiToken

pub enum ServiceAccountApiToken {
    Status(AccountNamedOpt),
    Generate {
        aopts: AccountCommonOpt,
        copt: CommonOpt,
        label: String,
        expiry: Option<String>,
    },
    Destroy {
        aopts: AccountCommonOpt,
        copt: CommonOpt,
        token_id: Uuid,
    },
}

Variants

Status(AccountNamedOpt)

Show the status of api tokens associated to this service account.

Generate

Fields

aopts: AccountCommonOpt

copt: CommonOpt

label: String

A string describing the token. This is not used to identify the token, it is only for human description of the tokens purpose.

expiry: Option<String>

An optional rfc3339 time of the format “YYYY-MM-DDTHH:MM:SS+TZ”, “2020-09-25T11:22:02+10:00”. After this time the api token will no longer be valid.

Generate a new api token for this service account.

Destroy

Fields

aopts: AccountCommonOpt

copt: CommonOpt

token_id: Uuid

The UUID of the token to destroy.

Destroy / revoke an api token from this service account. Access to the token is NOT required, only the tag/uuid of the token.