Enum kanidm_cli::ServiceAccountOpt

pub enum ServiceAccountOpt {
    Credential {
        commands: ServiceAccountCredential,
    },
    ApiToken {
        commands: ServiceAccountApiToken,
    },
    Posix {
        commands: ServiceAccountPosix,
    },
    Session {
        commands: AccountUserAuthToken,
    },
    Ssh {
        commands: AccountSsh,
    },
    List(CommonOpt),
    Get(AccountNamedOpt),
    Create(AccountCreateOpt),
    Update(ServiceAccountUpdateOpt),
    Delete(AccountNamedOpt),
    Validity {
        commands: AccountValidity,
    },
    IntoPerson(AccountNamedOpt),
}

Variants

Credential

Fields

commands: ServiceAccountCredential

Manage the generated password of this service account.

ApiToken

Fields

commands: ServiceAccountApiToken

Manage api tokens associated to this service account.

Posix

Fields

commands: ServiceAccountPosix

Manage posix extensions for this service account allowing access to unix/linux systems

Session

Fields

commands: AccountUserAuthToken

Manage sessions (user auth tokens) associated to this service account.

Ssh

Fields

commands: AccountSsh

Manage ssh public key’s associated to this person

List(CommonOpt)

List all service accounts

Get(AccountNamedOpt)

View a specific service account

Create(AccountCreateOpt)

Create a new service account

Update(ServiceAccountUpdateOpt)

Update a specific service account’s attributes

Delete(AccountNamedOpt)

Delete a service account

Validity

Fields

commands: AccountValidity

Manage a service account validity, such as expiry time (account lock/unlock)

IntoPerson(AccountNamedOpt)

Convert a service account into a person. This is used during the alpha.9 to alpha.10 migration to “fix up” accounts that were not previously marked as persons.

Implementations

impl ServiceAccountOpt

pub fn debug(&self) -> bool

pub async fn exec(&self)