Enum kanidm_cli::Oauth2Opt

pub enum Oauth2Opt {

    List(CommonOpt),
    Get(Named),
    CreateBasic(Oauth2BasicCreateOpt),
    UpdateScopeMap(Oauth2CreateScopeMapOpt),
    DeleteScopeMap(Oauth2DeleteScopeMapOpt),
    UpdateSupScopeMap(Oauth2CreateScopeMapOpt),
    DeleteSupScopeMap(Oauth2DeleteScopeMapOpt),
    ResetSecrets(Named),
    ShowBasicSecret(Named),
    Delete(Named),
    SetDisplayname(Oauth2SetDisplayname),
    SetName {
        nopt: Named,
        name: String,
    },
    SetLandingUrl {
        nopt: Named,
        url: String,
    },
    EnablePkce(Named),
    DisablePkce(Named),
    EnableLegacyCrypto(Named),
    DisableLegacyCrypto(Named),
    PreferShortUsername(Named),
    PreferSPNUsername(Named),

}

Variants

List(CommonOpt)

List all configured oauth2 resource servers

Get(Named)

Display a selected oauth2 resource server

CreateBasic(Oauth2BasicCreateOpt)

Create a new oauth2 resource server

UpdateScopeMap(Oauth2CreateScopeMapOpt)

Update or add a new mapping from a group to scopes that it provides to members

DeleteScopeMap(Oauth2DeleteScopeMapOpt)

Remove a mapping from groups to scopes

UpdateSupScopeMap(Oauth2CreateScopeMapOpt)

Update or add a new mapping from a group to scopes that it provides to members

DeleteSupScopeMap(Oauth2DeleteScopeMapOpt)

Remove a mapping from groups to scopes

ResetSecrets(Named)

Reset the secrets associated to this resource server

ShowBasicSecret(Named)

Show the associated basic secret for this resource server

Delete(Named)

Delete a oauth2 resource server

SetDisplayname(Oauth2SetDisplayname)

Set a new displayname for a resource server

SetName

Fields

nopt: Named

name: String

Set a new name for this resource server. You may need to update your integrated applications after this so that they continue to function correctly.

SetLandingUrl

Fields

nopt: Named

url: String

When redirecting from the Kanidm Apps Listing page, some linked applications may need to land on a specific page to trigger oauth2/oidc interactions.

EnablePkce(Named)

Enable PKCE on this oauth2 resource server. This defaults to being enabled.

DisablePkce(Named)

Disable PKCE on this oauth2 resource server to work around insecure clients that may not support it. You should request the client to enable PKCE!

EnableLegacyCrypto(Named)

Enable legacy signing crypto on this oauth2 resource server. This defaults to being disabled. You only need to enable this for openid clients that do not support modern crytopgraphic operations.

DisableLegacyCrypto(Named)

Disable legacy signing crypto on this oauth2 resource server. This is the default.

PreferShortUsername(Named)

Use the ‘name’ attribute instead of ‘spn’ for the preferred_username

PreferSPNUsername(Named)

Use the ‘spn’ attribute instead of ‘name’ for the preferred_username

Implementations

impl Oauth2Opt

pub fn debug(&self) -> bool

pub async fn exec(&self)